Rewrite and extend for n > 2
Before, this code assumed you were summing 2 digits. As opposed to simply
adding a 3-digit method, this rewrites the logic to accept any n and derive
the correct value from there. It does this using three queues, with a
base queue essentially getting processed and thrown into one of the two
others.

I originally was trying to do this with Array#each, but I couldn't find
a nice way to avoid processing the candidate numbers again. That's why I
moved to the queue system.