Day 3 tobaggan problem
Not too bad. Hardest part is figuring out how to deal with the x value.
Since the input is many times longer than wide, and you need to reach
the bottom, you need a way to index appropriately. This is solvable by
using the modulo operator.

You also have to be careful with going out of bounds while iterating.
The y value can't exceed the number of rows (or it could using modulo,
but that's not the exercise). Otherwise, just inspect the value you're
currently at and increment if it's a tree.