[Day 5] Working solution for both parts
This was a pretty easy one. I've been able to use the same base class
and subclass pattern for basically every one of these exercises, and
it worked nicely here to allow me to map the :id attribute accessor
across the seats and grab the max.

For part two, it was mostly a matter of finding something reliable to
use to determine when an increment of two happened. I chose to use the
index for that, but I imagine there are a few ways you could do it.