[Day 6] Working solution, both parts.
Ruby is so good. I usually prefer working in arrays versus strings, so
this accomplishes that using String#chars. The "unit" to work with was
groups; once you figure that out it's mostly a matter of batching the
group's forms together in a way such that you get the number of unique
letters among the group (part 1) or intersect the answers of each
group member to find the questions that were answers by everyone (part
2).