[Day 7] Code for mapping bag dependencies
This is only part of the solution for part 1. This code should result
in a hash, the keys of which are every color (in theory) in the input,
and the value of which is another hash with what bags can be contained
within that color bag. The keys of that hash are colors and the values
are a quantity. In theory this should be sufficient for the rest ...