[Day 8] Working solution for part 1 and 2
The class should track the accumulator and current instruction. It also
needs to track which instructions we've seen.

For part 2, I took all instructions visited between the first duplicate,
as that is when it would loop forever. I then iterated over those
instructions, swapping a jmp for a nop, until I found one that got to
the end. At that point, I just had to get the accumulator value.