[Day 12] Working part 1 & 2
This one was pretty easy. I have sliiiightly different implementations
for tracking an x,y location between the two parts, but it's fine. The
hard part is figuring out how to deal with the direction changes.

Once you recognize the math that is at play, it isn't so bad. I wrote
out the equivalent waypoint offsets in each direction and it was easy
to see you needed to inverse and possibly swap certain entries.