[Day 13] Working part 1, WIP part 2
This isn't efficient enough. There is something called the Chinese
Remainder Theorem that is apparently a possible solution. I'm not
doing that, but rather trying to make the largest jumps possible
between numbers I evaluate for the answer. This is still too slow and
did not produce an answer after 24 hours.

This works for simple lists so I think the solution is right, but I
need to look at the debug list and try to find what the "pattern" is.
Gotta be one.