[Day 14] Working part 2
Wow this was kind of a gross one. The main thing to recognize is that
you need 2^n variations of the address, where n is the number of
occurrences of an 'X'. At that point, it's a matter of constructing
all of the addresses. I did that by essentially creating every address
simultaneously, and at an 'X' I had some rather-icky logic that would
write 2^m 0's, then 2^m 1's. m was determined by which X I was on.

So for example, on the first X, you alternate writing 0 and 1 to an
address such that 2^0 = 1 times per digit. m would then increase, so
on the next X you would write 2^1 = 2 0's, then 2 1's, etc, until you
wrote to every in-progress address. This would continue until the last
X, where you would write n/2 0's and then n/2 1's.

I can't say this is my favorite code I have ever written, but it got
the job done and so ... we'll wrap it up from here.