[Day 15] Working part 1, 2
Part 2 was simple if your algorithm was computationally cheap.
Thankfully my solution was.

The main thing here is how to keep track of what numbers were said on
what turn. We don't care about the order that much, so we don't need
to keep a big list of what number was said each turn. The thing we
care about are what turns are relevant to a specific number. That
association made me think of a hash where the keys are numbers spoken
and the value is an array of turns.

From there it's as simple as subtracting the final two elements if the
array is greater than 1, and returning 0 if it isn't. You also need to
add one to the target because the turn counter starts at 1, not 0.
This is important for human-readability, but means the array, which
starts at 0, is artificially bigger according to the turn counter by
1.