[Day 16] Working part 1 solution
I cleaned this up a bunch. This one is tricky because there are nested
arrays and you have to be very careful about what condition you place
on them (all, any, none, etc.), plus there are both valid and invalid
tickets, so making sure you get the correct logic is critical.

Past that, it's a pretty simple "scan every ticket" exercise. The
biggest problem is likely that I separate out the tickets by valid and
invalid and then iterate over the invalid tickets again to find the
sum. Overall I'm okay with that because the "invalid sum" is kind of a
goofy-AoCism, and if we were writing a ticket scanner we would want
some logic for separating out valid and invalid tickets. So making
that the core feature and the invalid sum an extra seems okay.