[Day 16] Working part 2 solution
This adds a @positions class variable and two new methods that will
essentially calculate the list of possible fields for every index of a
ticket, as well as determine which field must correspond to which
position based on process of elimination.

The second method calculates a product of fields that contain
"departure" based on values in our ticket. This was the part 2 ask.