[Day 18] Dijkstra is my hero
The core problem was to do math that had parentheses in it and which
had strange precedence among operators. To get rid of parenthesis, I
used a shunting yard algorithm to convert the infix notation into
reverse polish (postfix) notation. This notation is nice because it is
unambiguous without parentheses.

Dealing with the precedence was interesting, but mostly okay. For the
most part you just treat everything equally or, in part 2, give
precedence to addition. This doesn't handle every operator, but that's
okay.

I love queues and I'm actually really pleased with this code. It was
fun to actually implement an algorithm from scratch, and even better
that it worked.