[Day 1] Rust 2020
This has poor performance since it's iterating over the same vector
once per iteration per iteration, but the input isn't too big and
we're learning so it's fine.

This was a good chance to work with Result, unwrap(), and expect(),
and I have a firmer handle on them now. The borrowing pattern is still
a little fuzzy, but kinda-sorta makes sense. I can at least reason
about what's specifically happening when I look over the code, even
though I'm relying a lot on the compiler to help me out right now.