[Day 2] Working solution for both parts
Ran into some fun little challenges with this and I imagine I could
optimize a lot. I was happy I thought to create a struct that would
represent each line in the input, and that worked basically really
well. It also allowed me to implement the two parts as methods on that
struct, which I thought was elegant.

This was the first time using the filter method in Rust and it's a bit
goofy to figure out. You can use iter() and collect() to move between
an iterator and a collection, which makes sense, but sometimes you end
up with these weird intermediary types that may or may not implement
all the methods you expect. This is a tricky part of rust but
navigating the docs is getting easier and I was able to figure it out.

I also had to do some casting with `as`, which is fine, but it took me
a bit to figure out exactly when in the chain of methods to do the
actual casting. Logical XOR (^) was also handy.